requirements.txt
Pillow-SIMD>=9.0.0
exifread>=3.0.0
watchdog>=2.1.0